Split Window into its own file
