+++ tempus_fugit_minecraft/main.py
@@ -1,625 +1,8 @@
 from __future__ import division
 
-import math
-import sys
-
+from pyglet import app
 from pyglet.gl import *
-from pyglet.window import key, mouse
-from tempus_fugit_minecraft.utilities import *
-from tempus_fugit_minecraft.model import Model
-
-if sys.version_info[0] >= 3:
-    xrange = range
-
-
-class Window(pyglet.window.Window):
-    """A window class for a game environment.
-
-    This class extends the `pyglet.window.Window` class and provides functionality
-    for a game environment. It handles player movement, collisions, rendering,
-    and user input.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(Window, self).__init__(*args, **kwargs)
-
-        # Whether the window exclusively captures the mouse.
-        self.exclusive = False
-
-        self.walking_speed = 5
-        self.FLYING_SPEED = 15
-
-        self.GRAVITY = 20.0
-        self.MAX_JUMP_HEIGHT = 1.0  # About the height of a block.
-        # To derive the formula for calculating jump speed, first solve
-        #    v_t = v_0 + a * t
-        # for the time at which you achieve maximum height, where a is the acceleration
-        # due to gravity and v_t = 0. This gives:
-        #    t = - v_0 / a
-        # Use t and the desired MAX_JUMP_HEIGHT to solve for v_0 (jump speed) in
-        #    s = s_0 + v_0 * t + (a * t^2) / 2
-        self.JUMP_SPEED = math.sqrt(2 * self.GRAVITY * self.MAX_JUMP_HEIGHT)
-        self.TERMINAL_VELOCITY = 50
-
-        self.PLAYER_HEIGHT = 2
-
-        # When flying gravity has no effect and speed is increased.
-        self.flying = False
-
-        # Strafing is moving lateral to the direction you are facing,
-        # e.g. moving to the left or right while continuing to face forward.
-        #
-        # First element is -1 when moving forward, 1 when moving back, and 0
-        # otherwise. The second element is -1 when moving left, 1 when moving
-        # right, and 0 otherwise.
-        self.strafe = [0, 0]
-
-        # Current (x, y, z) position in the world, specified with floats. Note
-        # that, perhaps unlike in math class, the y-axis is the vertical axis.
-        self.position = (0, 0, 0)
-
-        # First element is rotation of the player in the x-z plane (ground
-        # plane) measured from the z-axis down. The second is the rotation
-        # angle from the ground plane up. Rotation is in degrees.
-        #
-        # The vertical plane rotation ranges from -90 (looking straight down) to
-        # 90 (looking straight up). The horizontal rotation range is unbounded.
-        self.rotation = (0, 0)
-
-        # Which sector the player is currently in.
-        self.sector = None
-
-        # The crosshairs at the center of the screen.
-        self.reticle = None
-
-        # Velocity in the y (upward) direction.
-        self.dy = 0
-
-        # A list of blocks the player can place. Hit num keys to cycle.
-        self.inventory = [BRICK, GRASS, SAND]
-
-        # The current block the user can place. Hit num keys to cycle.
-        self.block = self.inventory[0]
-
-        # Convenience list of num keys.
-        self.num_keys = [
-            key._1, key._2, key._3, key._4, key._5,
-            key._6, key._7, key._8, key._9, key._0]
-
-        # Instance of the model that handles the world.
-        self.model = Model()
-
-        self.paused = False
-
-        # The label that is displayed in the top left of the canvas.
-        self.label = pyglet.text.Label(
-            text='',
-            font_name='Arial',
-            font_size=18,
-            x=10,
-            y=self.height - 10,
-            anchor_x='left',
-            anchor_y='top',
-            color=(0, 0, 0, 255)
-        )
-        self.pause_label = pyglet.text.Label(
-            text="Paused",
-            font_name="Arial",
-            font_size=36,
-            width=100,
-            height=30,
-            x=self.width // 2,
-            y=self.height // 2,
-            anchor_x="center",
-        )
-        self.resume_label = pyglet.text.Label(
-            text="Resume",
-            font_name="Arial",
-            font_size=18,
-            width=90,
-            height=35,
-            x=self.width // 2,
-            y=self.height // 2 - 45,
-            anchor_x="center",
-        )
-        self.quit_label = pyglet.text.Label(
-            text="Quit",
-            font_name="Arial",
-            font_size=18,
-            width=50,
-            height=35,
-            x=self.width // 2,
-            y=self.height // 2 - 90,
-            anchor_x="center",
-        )
-
-        # This call schedules the `update()` method to be called
-        # TICKS_PER_SEC. This is the main game event loop.
-        pyglet.clock.schedule_interval(self.update, 1.0 / TICKS_PER_SEC)
-
-    def set_exclusive_mouse(self, exclusive: bool) -> None:
-        """If `exclusive` is True, the game will capture the mouse, if False
-        the game will ignore the mouse.
-
-        Parameters
-        ----------
-        exclusive : bool
-            Whether the game will capture the mouse or not.
-        """
-        super(Window, self).set_exclusive_mouse(exclusive)
-        self.exclusive = exclusive
-
-    def get_sight_vector(self) -> tuple:
-        """Returns the current line of sight vector indicating the direction the player is looking."""
-        x, y = self.rotation
-        # y ranges from -90 to 90, or -pi/2 to pi/2, so m ranges from 0 to 1 and
-        # is 1 when looking ahead parallel to the ground and 0 when looking
-        # straight up or down.
-        m = math.cos(math.radians(y))
-        # dy ranges from -1 to 1 and is -1 when looking straight down and 1 when
-        # looking straight up.
-        dy = math.sin(math.radians(y))
-        dx = math.cos(math.radians(x - 90)) * m
-        dz = math.sin(math.radians(x - 90)) * m
-        return dx, dy, dz
-
-    def get_motion_vector(self) -> tuple:
-        """Returns the current motion vector indicating the velocity of the player.
-
-        Returns
-        -------
-        vector : tuple of len 3
-            Tuple containing the velocity in x, y, and z respectively.
-        """
-        if any(self.strafe):
-            x, y = self.rotation
-            strafe = math.degrees(math.atan2(*self.strafe))
-            y_angle = math.radians(y)
-            x_angle = math.radians(x + strafe)
-            if self.flying:
-                m = math.cos(y_angle)
-                dy = math.sin(y_angle)
-                if self.strafe[1]:
-                    # Moving left or right.
-                    dy = 0.0
-                    m = 1
-                if self.strafe[0] > 0:
-                    # Moving backwards.
-                    dy *= -1
-                # When you are flying up or down, you have less left and right
-                # motion.
-                dx = math.cos(x_angle) * m
-                dz = math.sin(x_angle) * m
-            else:
-                dy = 0.0
-                dx = math.cos(x_angle)
-                dz = math.sin(x_angle)
-        else:
-            dy = 0.0
-            dx = 0.0
-            dz = 0.0
-        return dx, dy, dz
-
-    def update(self, dt: float) -> None:
-        """This method is scheduled to be called repeatedly by the pyglet clock.
-
-        Parameters
-        ----------
-        dt : float
-            The change in time since the last call.
-        """
-        if self.paused:
-            return
-
-        self.model.process_queue()
-        sector = sectorize(self.position)
-        if sector != self.sector:
-            self.model.change_sectors(self.sector, sector)
-            if self.sector is None:
-                self.model.process_entire_queue()
-            self.sector = sector
-        m = 8
-        dt = min(dt, 0.2)
-        for _ in xrange(m):
-            self._update(dt / m)
-
-    def _update(self, dt: float) -> None:
-        """Private implementation of the `update()` method. This is where most
-        of the motion logic lives, along with gravity and collision detection.
-
-        Parameters
-        ----------
-        dt : float
-            The change in time since the last call.
-        """
-        # walking
-        speed = self.FLYING_SPEED if self.flying else self.walking_speed
-        d = dt * speed  # distance covered this tick.
-        dx, dy, dz = self.get_motion_vector()
-        # New position in space, before accounting for gravity.
-        dx, dy, dz = dx * d, dy * d, dz * d
-        # gravity
-        if not self.flying:
-            # Update your vertical speed: if you are falling, speed up until you
-            # hit terminal velocity; if you are jumping, slow down until you
-            # start falling.
-            self.dy -= dt * self.GRAVITY
-            self.dy = max(self.dy, -self.TERMINAL_VELOCITY)
-            dy += self.dy * dt
-        # collisions
-        x, y, z = self.position
-        x, y, z = self.collide((x + dx, y + dy, z + dz), self.PLAYER_HEIGHT)
-        self.position = (x, y, z)
-
-    def collide(self, position: tuple, height: int) -> tuple:
-        """Checks to see if the player at the given `position` and `height`
-        is colliding with any blocks in the world.
-
-        Parameters
-        ----------
-        position : tuple of len 3
-            The (x, y, z) position to check for collisions at.
-        height : int or float
-            The height of the player.
-
-        Returns
-        -------
-        position : tuple of len 3
-            The new position of the player taking into account collisions.
-        """
-        # How much overlap with a dimension of a surrounding block you need to
-        # have to count as a collision. If 0, touching terrain at all counts as
-        # a collision. If .49, you sink into the ground, as if walking through
-        # tall grass. If >= .5, you'll fall through the ground.
-        pad = 0.25
-        p = list(position)
-        np = normalize(position)
-        for face in FACES:  # check all surrounding blocks
-            for i in xrange(3):  # check each dimension independently
-                if not face[i]:
-                    continue
-                # How much overlap you have with this dimension.
-                d = (p[i] - np[i]) * face[i]
-                if d < pad:
-                    continue
-                for dy in xrange(height):  # check each height
-                    player_currnet_coords = list(np)
-                    player_currnet_coords[1] -= dy
-                    player_currnet_coords[i] += face[i]
-                    block_type = self.model.world.get(tuple(player_currnet_coords))
-                    if block_type is None or self.is_it_cloud_block(player_current_coords=tuple(player_currnet_coords)):
-                        continue
-                    p[i] -= (d - pad) * face[i]
-                    if face == (0, -1, 0) or face == (0, 1, 0):
-                        # You are colliding with the ground or ceiling, so stop
-                        # falling / rising.
-                        self.dy = 0
-                    break
-        return tuple(p)
-    
-
-    def on_mouse_press(self, x: int, y: int, button: int, modifiers: int) -> None:
-        """Called when a mouse button is pressed. See pyglet docs for button
-        and modifier mappings.
-
-        Parameters
-        ----------
-        x, y : int
-            The coordinates of the mouse click. Always center of the screen if
-            the mouse is captured.
-        button : int
-            Number representing mouse button that was clicked. 1 = left button,
-            4 = right button.
-        modifiers : int
-            Number representing any modifying keys that were pressed when the
-            mouse button was clicked.
-        """
-        if self.paused:
-            if self.within_label(x, y, self.resume_label):
-                self.resume_game()
-            elif self.within_label(x, y, self.quit_label):
-                self.close()
-        elif self.exclusive:
-            vector = self.get_sight_vector()
-            block, previous = self.model.hit_test(self.position, vector)
-            if (button == mouse.RIGHT) or ((button == mouse.LEFT) and (modifiers & key.MOD_CTRL)):
-                # ON OSX, control + left click = right click.
-                if previous:
-                    self.model.add_block(previous, self.block)
-            elif button == pyglet.window.mouse.LEFT and block:
-                texture = self.model.world[block]
-                if texture != STONE:
-                    self.model.remove_block(block)
-
-    @staticmethod
-    def within_label(x: int, y: int, label: pyglet.text.Label) -> bool:
-        """Returns True if the given (x, y) coordinates are within the given
-        label.
-
-        Parameters
-        ----------
-        x, y : int
-            The coordinates of the mouse click.
-
-        label : pyglet.text.Label
-            The label to check against.
-        """
-        x_within_range = label.x - label.width // 2 <= x <= label.x + label.width // 2
-        y_within_range = label.y <= y <= label.y + label.height // 2
-        return x_within_range and y_within_range
-
-    def on_mouse_motion(self, x: int, y: int, dx: int, dy: int) -> None:
-        """Called when the player moves the mouse.
-
-        Parameters
-        ----------
-        x, y : int
-            The coordinates of the mouse click. Always center of the screen if
-            the mouse is captured.
-        dx, dy : int
-            The movement of the mouse.
-        """
-        if self.paused:
-            if self.within_label(x, y, self.resume_label) or self.within_label(x, y, self.quit_label):
-                self.set_mouse_cursor(self.get_system_mouse_cursor(self.CURSOR_HAND))
-            else:
-                self.set_mouse_cursor(self.get_system_mouse_cursor(self.CURSOR_DEFAULT))
-
-            for label in [self.resume_label, self.quit_label]:
-                if self.within_label(x, y, label):
-                    label.color = (150, 150, 150, 255)  # grey
-                else:
-                    label.color = (255, 255, 255, 255)  # white
-                label.draw()
-
-        # Only rotate the camera if the mouse is captured.
-        if not self.exclusive or self.paused:
-            return
-
-        m = 0.15
-        x, y = self.rotation
-        x, y = x + dx * m, y + dy * m
-        y = max(-90, min(90, y))
-        self.rotation = (x, y)
-
-    def on_key_press(self, symbol: int, modifiers: int) -> None:
-        """Called when the player presses a key. See pyglet docs for key
-        mappings.
-
-        Parameters
-        ----------
-        symbol : int
-            Number representing the key that was pressed.
-        modifiers : int
-            Number representing any modifying keys that were pressed.
-        """
-        if symbol == key.ESCAPE:
-            if self.paused:
-                self.resume_game()
-            else:
-                self.pause_game()
-
-        if self.paused:
-            return
-
-        if symbol == key.W:
-            self.strafe[0] -= 1
-        elif symbol == key.S:
-            self.strafe[0] += 1
-        elif symbol == key.A:
-            self.strafe[1] -= 1
-        elif symbol == key.D:
-            self.strafe[1] += 1
-        elif symbol == key.SPACE:
-            if self.dy == 0:
-                self.dy = self.JUMP_SPEED
-        elif symbol == key.TAB:
-            self.flying = not self.flying
-        elif symbol in self.num_keys:
-            index = (symbol - self.num_keys[0]) % len(self.inventory)
-            self.block = self.inventory[index]
-        elif symbol == key.UP:
-            self.speed_up()
-        elif symbol == key.DOWN:
-            self.speed_down()
-
-    def speed_up(self) -> None:
-        """Increases the walking speed of the player."""
-        if self.walking_speed <= 15:
-            self.walking_speed += 5
-    def speed_down(self):
-        if self.walking_speed > 5:
-            self.walking_speed -= 5
-
-    def pause_game(self) -> None:
-        """Pauses the game and bring up the pause menu."""
-        self.paused = True
-        self.set_mouse_visible(True)
-        self.set_exclusive_mouse(False)
-
-    def resume_game(self) -> None:
-        """Resumes the game by restoring the game window to its original state."""
-        self.paused = False
-        self.set_exclusive_mouse(True)
-
-    def on_key_release(self, symbol: int, modifiers: int) -> None:
-        """Called when the player releases a key. See pyglet docs for key
-        mappings.
-
-        Parameters
-        ----------
-        symbol : int
-            Number representing the key that was pressed.
-        modifiers : int
-            Number representing any modifying keys that were pressed.
-        """
-        if symbol == key.W:
-            self.strafe[0] += 1
-        elif symbol == key.S:
-            self.strafe[0] -= 1
-        elif symbol == key.A:
-            self.strafe[1] += 1
-        elif symbol == key.D:
-            self.strafe[1] -= 1
-
-    def on_resize(self, width: int, height: int) -> None:
-        """Called when the window is resized to a new `width` and `height`.
-
-        Parameters
-        ----------
-        width : int
-            The new width of the window.
-        height : int
-            The new height of the window.
-        """
-        # label
-        self.label.y = height - 10
-        # reticle
-        if self.reticle:
-            self.reticle.delete()
-        x, y = self.width // 2, self.height // 2
-        n = 10
-        self.reticle = pyglet.graphics.vertex_list(
-            4,
-            ('v2i', (x - n, y, x + n, y, x, y - n, x, y + n))
-        )
-
-        if self.paused:
-            self.center_labels(width, height)
-
-    def center_labels(self, width: int, height: int) -> None:
-        """Center the labels when the window size changes.
-
-        Parameters
-        ----------
-        width : int
-            The new width of the window.
-        height : int
-            The new height of the window.
-        """
-        self.pause_label.x = self.resume_label.x = self.quit_label.x = width // 2
-        self.pause_label.y = height // 2
-        self.resume_label.y = height // 2 - 45
-        self.quit_label.y = height // 2 - 90
-
-    def set_2d(self) -> None:
-        """Configure OpenGL to draw in 2d."""
-        width, height = self.get_size()
-        glDisable(GL_DEPTH_TEST)
-        viewport = self.get_viewport_size()
-        glViewport(0, 0, max(1, viewport[0]), max(1, viewport[1]))
-        glMatrixMode(GL_PROJECTION)
-        glLoadIdentity()
-        glOrtho(0, max(1, width), 0, max(1, height), -1, 1)
-        glMatrixMode(GL_MODELVIEW)
-        glLoadIdentity()
-
-    def set_3d(self) -> None:
-        """Configure OpenGL to draw in 3d."""
-        width, height = self.get_size()
-        glEnable(GL_DEPTH_TEST)
-        viewport = self.get_viewport_size()
-        glViewport(0, 0, max(1, viewport[0]), max(1, viewport[1]))
-        glMatrixMode(GL_PROJECTION)
-        glLoadIdentity()
-        gluPerspective(65.0, width / float(height), 0.1, 60.0)
-        glMatrixMode(GL_MODELVIEW)
-        glLoadIdentity()
-        x, y = self.rotation
-        glRotatef(x, 0, 1, 0)
-        glRotatef(-y, math.cos(math.radians(x)), 0, math.sin(math.radians(x)))
-        x, y, z = self.position
-        glTranslatef(-x, -y, -z)
-
-    def on_draw(self):
-        """Called by pyglet to draw the canvas."""
-        self.clear()
-        self.set_3d()
-        glColor3d(1, 1, 1)
-        self.model.batch.draw()
-        self.draw_focused_block()
-        self.set_2d()
-        self.draw_label()
-        self.draw_reticle()
-
-        if self.paused:
-            self.draw_pause_menu()
-
-    def draw_pause_menu(self) -> None:
-        """Draws the components of the pause menu, including the background, the pause text, and the resume and quit
-        buttons.
-        """
-        glPushMatrix()
-        glLoadIdentity()
-        glMatrixMode(GL_PROJECTION)
-        glPushMatrix()
-        glLoadIdentity()
-        glOrtho(0, WINDOW_WIDTH, 0, WINDOW_HEIGHT, -1, 1)
-        glDisable(GL_DEPTH_TEST)
-
-        # Transparency
-        glEnable(GL_BLEND)
-        glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA)
-
-        background_color = (0, 0, 0, 0.8)
-        pyglet.graphics.draw(
-            4,
-            GL_QUADS,
-            ('v2i', (0, 0, WINDOW_WIDTH, 0, WINDOW_WIDTH, WINDOW_HEIGHT, 0, WINDOW_HEIGHT)),
-            ('c4f', background_color * 4)
-        )
-
-        glEnable(GL_DEPTH_TEST)
-        glPopMatrix()
-        glMatrixMode(GL_MODELVIEW)
-        glPopMatrix()
-
-        self.pause_label.draw()
-        self.resume_label.draw()
-        self.quit_label.draw()
-
-    def draw_focused_block(self) -> None:
-        """ Draw black edges around the block that is currently under the crosshairs."""
-        vector = self.get_sight_vector()
-        block = self.model.hit_test(self.position, vector)[0]
-        if block:
-            x, y, z = block
-            vertex_data = cube_vertices(x, y, z, 0.51)
-            glColor3d(0, 0, 0)
-            glPolygonMode(GL_FRONT_AND_BACK, GL_LINE)
-            pyglet.graphics.draw(24, GL_QUADS, ('v3f/static', vertex_data))
-            glPolygonMode(GL_FRONT_AND_BACK, GL_FILL)
-
-    def draw_label(self) -> None:
-        """Draw the label in the top left of the screen."""
-        x, y, z = self.position
-        self.label.text = '%02d (%.2f, %.2f, %.2f) %d / %d' % (
-            pyglet.clock.get_fps(), x, y, z,
-            len(self.model._shown), len(self.model.world))
-        self.label.draw()
-
-    def draw_reticle(self) -> None:
-        """Draw the crosshairs in the center of the screen."""
-        glColor3d(0, 0, 0)
-        self.reticle.draw(GL_LINES)
-    
-    
-    def is_it_cloud_block(self, player_current_coords):
-        """
-        check if the block at the given palyer_current_coords is a cloud block
-        
-        Input
-        -----
-            player_current_coords: current (x,y,z) corrdinates for the player
-        
-        Output
-        ------
-            True: if the coords corresponding to a cloud block.
-            False: otherwise.
-        """
-        block_type = self.model.world.get(player_current_coords)
-        return block_type in [LIGHT_CLOUD,DARK_CLOUD]
-
+from tempus_fugit_minecraft.window import Window, WINDOW_WIDTH, WINDOW_HEIGHT
 
 def setup_fog() -> None:
     """Configure the OpenGL fog properties."""
@@ -665,7 +48,7 @@ def main() -> None:
     # Hide the mouse cursor and prevent the mouse from leaving the window.
     window.set_exclusive_mouse(True)
     setup()
-    pyglet.app.run()
+    app.run()
 
 
 if __name__ == '__main__':
