+++ tempus_fugit_minecraft/utilities.py
@@ -66,8 +66,6 @@ def tex_coords(top: tuple, bottom: tuple, side: tuple) -> list:
     result.extend(side * 4)
     return result
 
-WINDOW_WIDTH = 800
-WINDOW_HEIGHT = 600
 TICKS_PER_SEC = 60
 GRASS = tex_coords((1, 0), (0, 1), (0, 0))
 SAND = tex_coords((1, 1), (1, 1), (1, 1))
